rename and generate
