+++ scripts/clog.inputs
@@ -15,8 +15,12 @@
 ../src/platform/datapath_winuser.c
 ../src/platform/datapath_raw_dpdk.c
 ../src/platform/datapath_raw_socket.c
-../src/platform/datapath_raw_xdp.c
-../src/platform/datapath_raw.c
+../src/platform/datapath_raw_socket_win.c
+../src/platform/datapath_raw_socket_linux.c
+../src/platform/datapath_raw_xdp_win.c
+../src/platform/datapath_raw_xdp_linux.c
+../src/platform/datapath_raw_win.c
+../src/platform/datapath_raw_linux.c
 ../src/platform/crypt_bcrypt.c
 ../src/platform/platform_winuser.c
 ../src/platform/toeplitz.c
