+++ src/platform/CMakeLists.txt
@@ -22,7 +22,7 @@ else()
     set(SOURCES ${SOURCES} inline.c platform_posix.c storage_posix.c cgroup.c)
     if(CX_PLATFORM STREQUAL "linux" AND NOT CMAKE_SYSTEM_NAME STREQUAL "FreeBSD")
         if(QUIC_USE_XDP)
-            set(SOURCES ${SOURCES} datapath_raw_unix.c datapath_raw_socket.c datapath_raw_socket_unix.c datapath_raw_xdp_unix.c)
+            set(SOURCES ${SOURCES} datapath_raw_linux.c datapath_raw_socket.c datapath_raw_socket_linux.c datapath_raw_xdp_linux.c)
         else()
             set(SOURCES ${SOURCES} datapath_epoll.c)
         endif()
